NetFlow/requirements.txt
Faker==18.4.0
influxdb_client==1.35.0
matplotlib==3.6.3
NetFlow==0.11.3
numpy==1.24.1
paho_mqtt==1.6.1
pandas==1.5.3
scikit_learn==1.2.2
scipy==1.10.1
silk==0.1
